:truck: Move HomeProductDetailSelectionView to a separate file
